second commit

+++ app/Http/Controllers/RegistrationController.php
@@ -6,7 +6,8 @@ use Illuminate\Http\Request;
 use Illuminate\Http\Response;
 
 use App\Http\Requests;
-
+use App\Meeting;
+use App\User;
 class RegistrationController extends Controller
 {
 
@@ -20,26 +21,31 @@ class RegistrationController extends Controller
         $meeting_id = $request->input('meeting_id');
         $user_id = $request->input('user_id');
 
-        $meeting=[
-            'title' => 'Title',
-            'description' => 'Description',
-            'time' => 'Time',
-            'view_meeting'=>[
-                'href'=>'api/v1/meeting/1',
-                'method'=>'GET'
+        $meeting=Meeting::findOrFail($meeting_id);
+        $user = User::findOrFail($user_id);
+
+        $message = [
+            'msg' => 'User is already registered for meeting',
+            'user' =>$user,
+            'meeting' =>$meeting,
+            'unregister' =>[
+                'href'=>'api/v1/meeting/registration/'.$meeting->id,
+                'method'=>'DELETE',
             ]
         ];
 
-        $user=[
-            'name'=>'Name'
-        ];
+        if($meeting->users()->where('users.id',$user->id)->first()){
+            return response()->json($message,404);
+        };
+
+        $user->meetings()->attach($meeting);
 
         $response = [
             'msg' =>'User registered for meeting',
             'meeting'=>$meeting,
             'user'=>$user,
             'unregister'=>[
-                'href'=>'api/v1/meeting/registration/1',
+                'href'=>'api/v1/meeting/registration/'.$meeting->id,
                 'method'=>'DELETE'
             ]
         ];
@@ -54,25 +60,14 @@ class RegistrationController extends Controller
 
     public function destroy($id)
     {
-        $meeting=[
-            'title' => 'Title',
-            'description' => 'Description',
-            'time' => 'Time',
-            'view_meeting'=>[
-                'href'=>'api/v1/meeting/1',
-                'method'=>'GET'
-            ]
-        ];
-
-        $user=[
-            'name'=>'Name'
-        ];
+        $meeting = Meeting::findOrFail($id);
+        $meeting->users()->detach();
 
         $response = [
             'msg' =>'User unregistered for meeting',
             'meeting'=>$meeting,
-            'user'=>$user,
-            'unregister'=>[
+            'user'=>'tbd',
+            'register'=>[
                 'href'=>'api/v1/meeting/registration',
                 'method'=>'POST',
                 'params'=>'user_id, meeting_id'
